brcm: Combine setup_environment and setup_u_boot into a single function.
u-boot does not need to set ARCH since it already knows what it is
from the Kconfig settings and board setup. This means we can
effectively use a single function for both main build and u-boot
setup.

The only difference between the two is setting up CROSS_COMPILE
for u-boot since we are currently building out of the tree.

Another improvement is not to keep adding to LD_LIBRARY_PATH but
just reset it any time the function is called.